backend/requirements.txt
fastapi
uvicorn[standard]
aiohttp
python-dotenv
sqlmodel
asyncpg
psycopg2-binary
google-generativeai
